Move handling of link context operation updates to appThunks.ts
